Debian packaging: Ignore missing dependency info for shared libraries
We use a custom tarball with libodbc 2.3.9 instead of the one on the system.
Debian packaging system will see us using the shared libraries from it
to build, thinking it creates a dependency users can't satisfy, and throw an error.

We tell it to ignore the issue, since the shared library from 2.3.9 is compatible
with the ones found on supported systems.