Bug #16590994 - ADO updates fail when recordset is adForwardOnly and recordset contains bit data.
This bug is repeatable VB6 logic when it generates the UPDATE query based on searchability of the column in a corresponding SELECT.
The driver converts BIT(N>1) to _binary and the comparison like " WHERE bit_col = _binary'....'" fails the query with the error message: Truncated incorrect DOUBLE value: "xxxxx".

Marking BIT(N>1) column as not searcheable will exclude it from being added in WHERE clause by ADO/VB6.
To check if column is searchable the user needs to call SQLColAttribute(hstmt, col_number, SQL_DESC_SEARCHABLE, ...);

Change-Id: I1a353b41c9f01ca643b2187ee8ed93c0be927fbb